+++ tests/testthat/test-cff_to_bibtex.R
@@ -12,10 +12,10 @@ test_that("Article to bibtex", {
     month = "January",
     keywords = c("Some", "simple", "keywords")
   )
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   x <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(x)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -40,10 +40,10 @@ test_that("Book to bibtex", {
     keywords = c("Two, keyword")
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -61,10 +61,10 @@ test_that("Booklet to bibtex", {
     keywords = "java"
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("InBook to bibtex with pages", {
@@ -89,10 +89,10 @@ test_that("InBook to bibtex with pages", {
     type = "Architecting Dependable Systems",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("InCollection to bibtex", {
@@ -108,10 +108,10 @@ test_that("InCollection to bibtex", {
     topic = "interrogatives;nl-semantics;polarity;"
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -133,10 +133,10 @@ test_that("InProceedings to bibtex", {
     address = "Murray Hill, New Jersey",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -151,10 +151,10 @@ test_that("Manual to bibtex", {
     howpublished = "CD-Rom",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -170,10 +170,10 @@ test_that("MastersThesis to bibtex", {
     note = "Forthcoming",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -187,10 +187,10 @@ test_that("PhdThesis to bibtex", {
     address = "Edinburgh",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -206,10 +206,10 @@ test_that("Proceedings to bibtex", {
     missinginfo = "pages",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("TechReport to bibtex", {
@@ -222,10 +222,10 @@ test_that("TechReport to bibtex", {
     address = "Pittsburgh, PA 15260",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("Unpublished to bibtex", {
@@ -237,10 +237,10 @@ test_that("Unpublished to bibtex", {
     missinginfo = "Date is  guess.",
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   bibparsed <- cff_parse_citation(bib)
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 
@@ -262,10 +262,10 @@ test_that("particle names", {
 
   expect_true(cff_validate(cffobj, verbose = FALSE))
 
-  expect_snapshot_output(bibparsed)
+  expect_snapshot(bibparsed)
 
   bib <- cff_to_bibtex(bibparsed)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("From plain cff with a citation", {
@@ -282,12 +282,12 @@ test_that("From plain cff with a citation", {
   s$`preferred-citation`$editors <- list(cff_parse_person("A name"))
 
   bib <- cff_to_bibtex(s)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("From plain cff", {
-  bib <- cff_to_bibtex(cff())
-  expect_snapshot_output(toBibtex(bib))
+  expect_message(bib <- cff_to_bibtex(cff()))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("From file", {
@@ -296,7 +296,7 @@ test_that("From file", {
   )
 
   bib <- cff_to_bibtex(file)
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
 })
 
 test_that("NULL", {
@@ -311,8 +311,8 @@ test_that("Test anonymous", {
   )
 
 
-  back <- cff_to_bibtex(cff_parse_citation(bib))
-  expect_snapshot_output(toBibtex(back))
+  expect_message(back <- cff_to_bibtex(cff_parse_citation(bib)))
+  expect_snapshot(toBibtex(back))
 
 
   bib <- bibentry("manual",
@@ -320,16 +320,16 @@ test_that("Test anonymous", {
   )
 
 
-  back <- cff_to_bibtex(cff_parse_citation(bib))
-  expect_snapshot_output(toBibtex(back))
+  expect_message(back <- cff_to_bibtex(cff_parse_citation(bib)))
+  expect_snapshot(toBibtex(back))
 
   bib <- bibentry("misc",
     title = "A misc"
   )
 
 
-  back <- cff_to_bibtex(cff_parse_citation(bib))
-  expect_snapshot_output(toBibtex(back))
+  expect_message(back <- cff_to_bibtex(cff_parse_citation(bib)))
+  expect_snapshot(toBibtex(back))
 
   bib <- bibentry("proceedings",
     title = "proceedings",
@@ -337,8 +337,8 @@ test_that("Test anonymous", {
   )
 
 
-  back <- cff_to_bibtex(cff_parse_citation(bib))
-  expect_snapshot_output(toBibtex(back))
+  expect_silent(back <- cff_to_bibtex(cff_parse_citation(bib)))
+  expect_snapshot(toBibtex(back))
 })
 
 test_that("Fallback month", {
@@ -350,14 +350,14 @@ test_that("Fallback month", {
     date = "1678-04-23"
   )
 
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   x <- cff_parse_citation(bib)
 
   # Delete here the month
   x$month <- NULL
 
   bibback <- cff_to_bibtex(x)
-  expect_snapshot_output(toBibtex(bibback))
+  expect_snapshot(toBibtex(bibback))
 })
 
 
@@ -386,10 +386,75 @@ test_that("Test BibLateX entry", {
     issn = "0097-8493",
     url = "http://www.ctan.org"
   )
-  expect_snapshot_output(toBibtex(bib))
+  expect_snapshot(toBibtex(bib))
   x <- cff_parse_citation(bib)
 
 
   parsed <- cff_to_bibtex(x)
-  expect_snapshot_output(toBibtex(parsed))
+  expect_snapshot(toBibtex(parsed))
+})
+
+
+test_that("Test Fallback year", {
+  x <- cff()
+
+  expect_message(msg <- cff_to_bibtex(x))
+
+  expect_snapshot(toBibtex(msg))
+
+
+  x$`date-released` <- "2020-01-01"
+
+  expect_true(cff_validate(x, verbose = FALSE))
+
+  parsed <- cff_to_bibtex(x)
+
+  expect_snapshot(toBibtex(parsed))
+})
+
+test_that("Errors", {
+  expect_silent(b <- cff_to_bibtex("testthat"))
+  expect_s3_class(b, "bibentry")
+  expect_error(cff_to_bibtex("testthat", what = "aa"))
+})
+
+test_that("From package", {
+  skip_if_not_installed("rmarkdown")
+
+  base <- cff_to_bibtex("rmarkdown")
+
+  expect_s3_class(base, "bibentry")
+
+  expect_length(base, 1)
+
+  refs <- cff_to_bibtex("rmarkdown", "references")
+  expect_s3_class(refs, "bibentry")
+
+  expect_gte(length(refs), 1)
+
+  all <- cff_to_bibtex("rmarkdown", "all")
+  expect_s3_class(all, "bibentry")
+
+  expect_length(all, length(base) + length(refs))
+})
+
+test_that("NULL references", {
+  basic <- cff()
+
+  expect_null(cff_to_bibtex(basic, "references"))
+
+  # Test all
+  expect_message(l <- cff_to_bibtex(basic, "all"))
+  expect_length(l, 1)
+})
+
+
+test_that("From CITATION.cff", {
+  p <- system.file("examples/smith-et-al.cff", package = "cffr")
+
+  base <- cff_to_bibtex(p)
+
+  expect_s3_class(base, "bibentry")
+
+  expect_length(base, 1)
 })