+++ NEWS.md
@@ -1,3 +1,32 @@
+# cffr 0.5.0
+
+## Lifecycle
+
+-   Function `cff_to_bibtex()` and `cff_extract_to_bibtex()` have been merged,
+    the latter (`cff_extract_to_bibtex()`) is now soft-deprecated.
+
+## New features
+
+-   New function `write_citation()` that can generate a inst/CITATION file from
+    a `cff` object (#51).
+
+## Enhancements
+
+-   Additional authors of a **R** package can be now included based on the role
+    on the DESCRIPTION file, via the parameter `authors_roles` (#49).
+
+-   New message interface based on [**cli**](https://cli.r-lib.org/)
+    capabilities.
+
+-   Now the (invisible) result of `cff_validate()` includes the summary of
+    errors (if any) as an `attributes()` named `"error"`, as
+    `jsonvalidate::json_validate()` does when `verbose = TRUE`:
+
+    > Be verbose? If `TRUE`, then an attribute `"errors"` will list validation
+    > failures as a data.frame
+
+-   Improvements in the extraction of `` `date-release` ``.
+
 # cffr 0.4.1
 
 -   Replace **crayon** dependency for **cli** (#46).